Create generic Frame meta class
This creates a frame meta class which acts as a parent class to
implementations for each file type. This approach is used to maintain
consistency between file readers and to make getting attributes simpler.

Along with the metaclass I have an implementation of it with gsdFrame,
which is a wrapper around the gsd.snapshot class. The instances of this
class replace the existing implementation.