Include max_relaxation function
For some dynamics quantities, I am currently thinking about the
non-gaussian, the natural relaxation time, or quantity which best
represents the characteristics of the simulation is the time at which
the maximum occurs. Additionally the value of the maximum can be useful
so I have also included that as a return value.