Properly parameterise testing of frames.
The tests for the frames was not properly doing to parameterisation,
which is somewhat of a limitation of pytest itself. Rather than try and
work around the problem, I am now using the
[pytest-lazy-fixture](https://github.com/TvoroG/pytest-lazy-fixture)
module which allows for the fixtures to be used as parameters to other
fixtures.

This change makes the testing of the frames much simpler.