Handle legacy representation of molecules in gsd files
This updates the gsdFrame class to handle the case where all the
particles of a molecule are included in the input file. I am no longer
using this storage format, however there are older files which are. By
putting the logic in the gsdFrame class, I am making the interface for
any of the filetypes I have the same.

This should remove many of the issues I have been having with the old
file representation.