Add x,y,z position properties to Frame metaclass
This adds the x_position, y_position and z_position property functions
to the Frame metaclass, and each of the child classes.

These methods have been included for simplicity, since extracting only
the parameters of the cardinal directions is a relatively common
operation I have made sure that they are present.