Create custom bokeh glyph for trimer
This is a custom glyph for bokeh which uses JavaScript to draw a trimer
molecule from orientation, and positional data. This makes the
plotting of the molecules significantly simpler, not having to work out
where the positions of each of the circles are located.

In creating a custom glyph I am also able to control the drawing of the
lines, enabling the use of a bordering line surrounding the shape.