Refactor drawing of configuration to use trimer glyph
This is a significant refactoring of the code to use the custom glyph
introduced in the previous commit to plot the trimer molecule. This also
uses the Frame class to access the data from each configuration, which
generalises the plotting to any class implementing that interface.

In refactoring the code I have also used an updated version of bokeh,
which implements more functionality for defining the colours of the
figures. There is also an HSL class implemented, allowing the
specification of the colours directly from bokeh rather than having to
specify the colour maps manually.