Major rewrite of interactive_config
This restructures the bokeh server application into a class, since it
constitutes a state. I am actually not entirely sure this shouldn't be
some nested functions, however that change in architecture would be a
relatively straightforward change.

By reformulating this as a class, which really means a collection of
functions operating on a state. I have hopefully reduced some of the
complexity of what is going on, in that everything is nicely labelled.