TST Use TemporaryDirectory for output directory
When running the sdanalysis_test I was running into an error where the
data couldn't be appended to the existing data. This was a result of
using the same directory for every test without removing the data from
the old tests. This old data is now removed automatically using the
TemporaryDirectory class preventing this error from occurring again and
also cleaning up my working directory.