ENH Support plotting circles and glyphs
In bokeh, plotting the trimer glyph turns out to significantly slower
than circles; likely due to mathematical operations being
significantly optimised in python. An additional factor is circles
support openGL rendering, which provides additional rendering
efficiencies. To enable both performant and pretty visualisations I have
added support for plotting the Trimer molecule as a glyph, or as a
collection of circles as previously.