ENH Improve performance of calculating colour
This is a significant improvement to the performance of the colouring of
particles. All colours are pre-calculated on the loading of the module
which gives 360 light and 360 dark colours. The act of choosing the
colouring is then converting the orientation to an integer in degrees
which is the colouring of the particle. This can be very simply achieved
using the fancy indexing of numpy arrays.