REL Include important non-python files in manifest.in
This is an issue with the packaging in not including all the required
files for installation. It appears that the pyproject.toml file was not
included since the installation from sdist still didn't proceed as
expected. Additionally the trimer.ts file wasn't included since it isn't
a python file, however it is required.