ENH Use dataclasses for frame classes
The dataclass code generator uses all the best practices when creating a
class; I don't. To improve the quality of the code and functionality of
the Frame classes, using dataclasses seems like a really good idea.
At this stage I am using a back ported version, however this will be
built into python 3.7 which is even more reason to use this feature.