BUG Make copy of array when converting types
This fixes a bug in the computation of the translational displacement
when the values that form the box are a float64 type instead of a
float32. The issue was in the conversion of the type without copying,
leaving an array that was inaccessible through cython.

I believe this issue came about as a result of a change in the gsd file
specification, so this was not present in the older files I am using for
testing. I have included a specific test for this issue so it shouldn't
appear again.