ENH Use attrs for params class
This is a simplified version of the params class in the
statdyn-simulation repository using the attrs class. I am using attrs
over dataclass since I have found that it works significantly better for
what I am trying to do when I was working with the statdyn-simulation
repository.

This also includes attrs as a dependency in all the different deployment
methods.