ENH Move from dataclasses to attrs for class generation
Using attrs is consistent with the SimulationParams class, which also
removes the dataclasses dependency. I feel this is also a usability
improvement, with attrs just working for me far more than dataclasses.

In this move to update the classes I have also updated the names of the
classes to start with a capital for consistency and adherence to pep8.