BUG Create linear_steps attribute in SimulationParams
The SimulationParams class was missing an attribute required for the
complete use of the class. The linear_steps attribute is the number of
steps between exponential increases in the step_size. In almost all
cases this is 100, however some of the older datasets used a value of
10. To enable the processing of these older files I have added a this
attribute to the SimulationParams class and an option on the command
line.