BLD Move installation of sdanalysis from pipfile to makefile
The installation of this repository, sdanalysis, is failing on travis
since the prerequisites, numpy and cython, are not installed. To remedy
this I am moving the installation of sdanalysis to it's own step in the
makefile and in the docker container. This will prevent the issues I am
currently seeing.