ENH Only draw molecules as a collection of circles
The drawing of the complicated trimer shape, while it looked nice, it
took too long to draw making the interaction experience not so
desirable. Additionally I hadn't implemented the 'hitbox' of the shape
so there was no way to perform mouse-over actions.

The drawing of the trimer molecule, and all other molecules is now as a
collection of circles.