ENH Compute dynamics for a list of files
This uses the python multiprocessing module to process a collection of
files in parallel, while using an alternate thread to write the files.
There is no support for parallel writing of HDF5 files and alternate
methods don't work, so this is the only method of parallelism which will
work given the constraints I am working in.

It turns out that creating parallel threads for processing is rather
simple once the appropriate functions are set up.