BUG Handle NaN values by filtering not averaging
I made the assessment that for the Molecular Relaxations, an initial
frame which has any NaN values for the relaxation time is invalid, since
it has not had time to properly equilibrate. In handling these values I
was taking the mean over the configuration, which when there were any
NaN values, the resulting mean would be NaN. However, also calculating
the harmonic mean caused issues, since I needed the values for all the
molecules and not just for each frame.

To fix the issue I am now filtering for the NaN values before
calculating both the mean and the harmonic mean, ensuring both these
quantities have valid values.