fix: Handle errors when running with no crystal value
Where there is no value for the crystal, the interactive configuration
didn't handle at all, unwinding to a stack trace. This provides
support for that case with additional error handling, particularly for
values that don't exist. However this is at the expense of handling the
crystal case, which will need to be added at a future point in time.