style: Allow any ordering of arguments in a conditional
A conditional where there are multiple checks, i.e. 1 < val and val < 3
is much easier to read where all the values are ordered as per the
number line. There was a rule in pylint which wants constants last, so I
have turned the rule off for the sake of clarity in the interpretation
of the values.