fix: Handle diffusion constant where there is no relaxation
In calculating the diffusion constant I am interested in the linear
region of the dynamics, i.e. where the msd follows a power law of 1. For
this I check that the msd is between 2 and 100, which is typically the
range for the onset of the linear region. Where there were no points in
this region the analysis would crash outputting no results. This commit
checks there are sufficient values and returning a Null value if not.