chore: Remove scikit learn models from package
The scikit learn models are now removed from the package since there are
issues with requiring the exact version of sklearn. An additional issue
is that there is no simple tooling for regenerating and updating these
models, which is compounded by the issue that they only work for the
molecules I am working with, having no generality. The final reason for
removal is that the molecular definition has been modified the models
are not really valid any more.

BREAKING CHANGES

- Machine learning models removed