fix: Disable calculation of displacement across images
It turns out that the output files from hoomd are doing unexpected
things with the image a particle is occupying. There are occasions where
a particle close to the edge crosses the periodic boundary yet the image
is not updated. This can occur for both one of the edges, or it has also
occurred on both edges. This is really problematic for the MSD and
similar calculations and so has been disabled.