docs: Update modules in project documentation
The documentation of the project has been neglected for a very long
time, with many changes to the structure of the project since it was
last touched. These changes bring the structure of the documentation in
line with that of the project, although there is still much to do in the
way of making the documentation usable.