feat: Implement calculating the intermediate scattering function
The intermediate_scattering_function is a standard measure of the
structural relaxation of a liquid. I have not been measuring this
quantity to date, instead opting for a simpler version. However, I think
there are modifications required for the simpler version so this is a
check. There should be no changes required to properly get the values
from this function.

This changes the default number of dimensions in the Dynamics class from
3 to 2.