feat: Add progress bars to processing files
When processing a file, there are now progress bars indicating how much
of the processing has been completed. This calculates the number of
frames which have been processed, which particularly on the final
configurations is a bit of an underestimate since there are multiple
comparisons for each frame. However it does provide an interactive and
updating view of progress and an estimate of time remaining.

This is an overwhelmingly nice improvement since it allows me to see how
much longer a processing task is taking and additionally there is scope
to troubleshoot why something is going slow since I now have a
reference.