feat: Raise exceptions in preference of assertions
Rather than using assertions to make checks for issues with the inputs
to functions, this instead uses the exceptions RuntimeError and
ValueError depending on the issue which has been found. The ValueError
is where the user input is incorrect while a RuntimeError is where there
is a state which is not correct somewhere within the function.