feat: Command line options for serial and parallel processing
Processing files in parallel has the issue that should a single file
have issues then everything needs to be recomputed, which can be a
lengthy process. While in the serial processing mode, only a single file
needs re-processing should it have errors. This allows for the choice to
be made at runtime to suit the particular use case.