refactor: Split processing of gsd files
The process_gsd function was getting really long, handling many cases.
This splits the function into separate parts, each handling a specific
case. This makes good use of the yield from construct as it is an
iterator.

The split handles the different types of gsd files I am processing,
those with a linear sequence of steps, and those with an exponential
sequence.