fix: Use threshold_relaxation for exponential relaxations
The curve fitting for the exponential relaxations is not working
properly, giving results which are clearly wrong. Rather than
attempting to get it fixed or have a whole bunch of data cleaning, I am
just switching to the simpler threshold_relaxation, which looks for the
first time for which the value dips below the specified threshold.