refactor: Simplify LastMolecularRelaxation
This refactors the add method of the LastMolecularRelaxation class to
operate as a sate machine, with the focus on moving from the starting
state being 0 to the final state 2. It defines the different ways to
move between the states. As part of developing this code a simple linear
algorithm was written which has been left in the comments.

This simplifies the code significantly and the idea is to help find a
bug which is causing unusual values for the relaxation time.