fix: Pass correct displacements to relaxation instances
This ensures the correct displacements, whether they be translational or
rotational are passed to the method calculating the relaxation values.
Previously I had been using the name to determine the type of
relaxation, however I changed the names of the functions to better match
those potentially for publication which broke the conditional. This adds
a property of the instance determining which values are added.