fix: Handle None value of models
The default value of models was a list, which apart from being a
terrible idea since the list is the same for all instantiations. When
running on the command line the default value is None, so this changes
the default to None and adds the logic to handle the None case.