feat: Allow specification of factors in configuration
The use of the configuration is often done as a comparison of many
different configurations. Because of this is is particularly useful to
have the same colourscheme between all the figures. This commit adds
support for specifying all the factors to a configuration so that the
colourscheme is consistent with other figures.