refactor: Remove threaded processing implementation
The threading is causing more headaches than it helps, so this is being
removed from the package. It is much simpler to have multiple threads
running the processing into different files, then combining all the
files later, rather than trying to do everything in a single step and
having to re-run everything when a single part breaks.

BREAKING:
    - Remove parallel_process_files