refactor: Calculate dynamics conditional on wavenumber
The wavenumber is a value which is used as a distance metric for the
dynamics quantities. For this value to be useful it needs to be passed
in by the user. Rather than trying to compute it automatically, when
there is no wavenumber value given, the computation will ignore any
dynamics quantities which require the wavenumber.