feat: Command line option to calculate scattering function
The intermediate scattering function takes somewhat longer to calculate
than the rest of the dynamics quantities so it is not enabled by
default, however it is useful to be able to compute it. This enables the
command line flag `--scattering-function` which allows the calculation
of the intermediate scattering function to be specified at runtime.