fix: Calculate structural relaxation correctly
This fixes some issues with the calculation of the structural
relaxation. The main issue being the incorrect order of the rotations
when creating the rotation quaternion, and additionally not negating the
initial particle positions from the updated positions.

There are also now a range of tests which verify the calculations are
correct.

Fixes #155